Update code : after review by Thanh
