+++ src/main/java/com/automatedtest/sample/utils/Common.java
@@ -1,5 +1,9 @@
 package com.automatedtest.sample.utils;
 
+import java.io.File;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.nio.file.Paths;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -11,5 +15,12 @@ public class Common {
             return matcher.group(groupIdx);
         }
         return null;
+    }
+
+    public static String getResourceAbsolutePath(Object object, String filePath) throws URISyntaxException {
+        URL res = object.getClass().getClassLoader().getResource(filePath);
+        File file = Paths.get(res.toURI()).toFile();
+        String absolutePath = file.getAbsolutePath();
+        return absolutePath;
     }
 }
