+++ src/test/java/com/automatedtest/sample/definitions/SearchImageSteps.java
@@ -1,18 +1,14 @@
 package com.automatedtest.sample.definitions;
 
-import com.automatedtest.sample.pages.ImageTabOfSearchResultPage;
 import com.automatedtest.sample.pages.SearchResultPage;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.When;
 
-public class SearchImageByDropLink {
+public class SearchImageSteps {
     private final SearchResultPage searchResultPage;
-    private final ImageTabOfSearchResultPage imageTabOfSearchResultPage;
 
-
-    public SearchImageByDropLink() {
+    public SearchImageSteps() {
         this.searchResultPage = new SearchResultPage();
-        this.imageTabOfSearchResultPage = new ImageTabOfSearchResultPage();
     }
 
     @And("I change to image tab")
@@ -21,7 +17,7 @@ public class SearchImageByDropLink {
     }
 
     @When("I drag and drop first image result to search image")
-    public void aUserDragAndDropFirstImageResultToSearchImage() {
-        this.imageTabOfSearchResultPage.dragNDropFirstImageToSearch();
+    public void aUserDragAndDropFirstImageResultToSearchImage() throws Exception {
+        this.searchResultPage.dragNDropFirstImageToSearch();
     }
 }
