+++ src/test/java/com/automatedtest/sample/definitions/SearchResultPageSteps.java
@@ -1,18 +1,26 @@
 package com.automatedtest.sample.definitions;
 
+import com.automatedtest.sample.pages.ImageSearchPage;
 import com.automatedtest.sample.pages.SearchResultPage;
 import com.automatedtest.sample.pages.YoutubePage;
+import io.cucumber.java.en.And;
+import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
 import org.junit.Assert;
+
+import java.net.URISyntaxException;
 
 public class SearchResultPageSteps {
 
     private final SearchResultPage searchResultPage;
     private final YoutubePage youtubePage;
+    private final ImageSearchPage imageSearchPage;
 
     public SearchResultPageSteps() {
         this.searchResultPage = new SearchResultPage();
         this.youtubePage = new YoutubePage();
+        this.imageSearchPage = new ImageSearchPage();
     }
 
     @Then("^\"([^\"]*)\" is displayed in all the results$")
@@ -32,5 +40,26 @@ public class SearchResultPageSteps {
         String videoName = this.searchResultPage.openVideoAndGetName();
         Assert.assertEquals("First video of Google search result should be opened correctly",
                 videoName, this.youtubePage.getVideoTitle());
+    }
+
+    @Given("I'm on Google Image Search page")
+    public void aUserOnGoogleImageSearchPage() {
+        this.imageSearchPage.goToImageSearchPage();
+    }
+
+    @When("I search for image with file path {string}")
+    public void aUserSearchForImageWithFilePath(String filePath) throws URISyntaxException {
+        this.imageSearchPage.uploadImage(filePath);
+    }
+
+    @Then("Image with file path {string} should be uploaded on search page")
+    public void imageShouldBeUploadedOnSearchPage(String filePath) {
+        String imageName = this.imageSearchPage.getImageSearchName();
+        Assert.assertTrue("Image should be uploaded on search page", filePath.contains(imageName));
+    }
+
+    @And("Search result page should have image result")
+    public void searchResultPageShouldHaveImageResult() {
+        Assert.assertTrue("Search result page should have image result", this.imageSearchPage.getListImageResult().size() > 0);
     }
 }
