+++ src/test/resources/com/automatedtest/sample/ImageSearch.feature
@@ -4,11 +4,10 @@ Feature: SearchByImage
   Scenario Outline: Verify Google Image search
     Given I'm on Google Image Search page
     When I search for image with file path "<filePath>"
-    Then Image "<filePath>" should be uploaded on search page
+    Then Image with file path "<filePath>" should be uploaded on search page
     And Search result page should have image result
 
-
     Examples:
-      | filePath           |
+      | filePath            |
       | \image\house.jpg    |
